Review comment fix.
